+++ pysat/tests/test_instruments.py
@@ -81,9 +81,6 @@ def generate_instrument_list(instrument_names=[], package='pysat.instruments'):
                         if inst._test_download:
                             if not travis_skip:
                                 instrument_download.append(inst)
-                            else:
-                                print(' '.join(['Skipping', name, sat_id, tag,
-                                                'on Travis CI']))
                         else:
                             # we don't want to test download for this combo
                             print(' '.join(['Excluding', name,
@@ -127,9 +124,6 @@ class TestInstrumentsAll():
 
         for sat_id in module.sat_ids.keys():
             for tag in module.sat_ids[sat_id]:
-                print(' '.join(('Checking pysat.Instrument',
-                                'instantiation for module:', name,
-                                'tag:', tag, 'sat id:', sat_id)))
                 inst = pysat.Instrument(inst_module=module, tag=tag,
                                         sat_id=sat_id)
                 assert isinstance(inst, pysat.Instrument)
@@ -191,22 +185,17 @@ class TestInstrumentsDownload():
                         inst.platform, inst.name, inst.tag, inst.sat_id)))
         if len(inst.files.files) > 0:
             inst.clean_level = clean_level
-            inst.data = pds.DataFrame([0])
+            target = 'Fake Data to be cleared'
+            inst.data = [target]
             start = inst._test_dates[inst.sat_id][inst.tag]
             inst.load(date=start)
+            # Make sure fake data is cleared
+            assert target not in inst.data
+            # If cleaning not used, something should be in the file
+            # Not used for other levels since cleaning may remove all data
             if clean_level == "none":
-                # Something should be in the file
                 assert not inst.empty
-            else:
-                # Alternate check since cleaning may remove all data
-                try:
-                    assert inst.data != pds.DataFrame([0])
-                except (ValueError, AssertionError):
-                    # if objects cannot be compared, not the same
-                    # ValueError when wrong pandas object is here
-                    # AssertionaError if xarray
-                    assert True
-
+            # For last parametrized clean_level, remove files
             if clean_level == "clean":
                 remove_files(inst)
         else: