+++ pysat/tests/test_custom.py
@@ -1,5 +1,4 @@
 import numpy as np
-import warnings
 
 from nose.tools import raises
 import pandas as pds
@@ -17,9 +16,9 @@ class TestBasics():
         """Runs after every method to clean up previous testing."""
         del self.testInst
 
-    def add(self, function, kind='add', at_pos='end', *args, **kwargs):
-        '''Adds a function to the object's custom queue'''
-        self.testInst.custom.add(function, kind, at_pos, *args, **kwargs)
+    def attach(self, function, kind='add', at_pos='end', *args, **kwargs):
+        """Adds a function to the object's custom queue"""
+        self.testInst.custom.attach(function, kind, at_pos, *args, **kwargs)
 
     @raises(ValueError)
     def test_single_modifying_custom_function(self):
@@ -31,7 +30,7 @@ class TestBasics():
             inst.data['doubleMLT'] = 2.0 * inst.data.mlt
             return 5.0 * inst.data['mlt']
 
-        self.testInst.custom.add(custom1, 'modify')
+        self.testInst.custom.attach(custom1, 'modify')
         self.testInst.load(2009, 1)
 
     def test_single_adding_custom_function(self):
@@ -43,7 +42,7 @@ class TestBasics():
             d.name = 'doubleMLT'
             return d
 
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert (self.testInst['doubleMLT'].values == 2.0 *
                 self.testInst['mlt'].values).all()
@@ -59,7 +58,7 @@ class TestBasics():
             print(new_index)
             return d
 
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         ans = (self.testInst['doubleMLT'].isnull()).all()
         if self.testInst.pandas_format:
@@ -78,7 +77,7 @@ class TestBasics():
             inst['mlt'] = 0.
             return inst.data.doubleMLT
 
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert (self.testInst.data['doubleMLT'] == 2.0 *
                 self.testInst['mlt']).all()
@@ -89,7 +88,7 @@ class TestBasics():
         """
         def custom1(inst):
             return ('doubleMLT', 2.0 * inst.data.mlt.values)
-        self.testInst.custom.add(custom1, 'add')
+        self.testInst.custom.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         print(self.testInst['doubleMLT'])
         print(2.0 * self.testInst['mlt'])
@@ -102,7 +101,7 @@ class TestBasics():
         def custom1(inst):
             return (['doubleMLT', 'tripleMLT'], [2.0 * inst.data.mlt.values,
                                                  3.0 * inst.data.mlt.values])
-        self.testInst.custom.add(custom1, 'add')
+        self.testInst.custom.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert (self.testInst.data['doubleMLT'] == 2.0 *
                 self.testInst['mlt']).all()
@@ -116,7 +115,7 @@ class TestBasics():
         """
         def custom1(inst):
             return ('doubleMLT', 2.0 * inst.data.mlt.values[0:-5])
-        self.testInst.custom.add(custom1, 'add')
+        self.testInst.custom.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         if not self.testInst.pandas_format:
             print("Warning! Xarray doesn't enforce the same number of " +
@@ -130,7 +129,7 @@ class TestBasics():
         """
         def custom1(inst):
             return ('doubleMLT', np.arange(2.0 * len(inst.data.mlt)))
-        self.testInst.custom.add(custom1, 'add')
+        self.testInst.custom.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         if not self.testInst.pandas_format:
             print("Warning! Xarray doesn't enforce the same number of " +
@@ -143,7 +142,7 @@ class TestBasics():
                                    'tripleMLT': inst.data.mlt * 3},
                                   index=inst.index)
             return out
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert (self.testInst.data['doubleMLT'] == 2.0 *
                 self.testInst['mlt']).all()
@@ -158,7 +157,7 @@ class TestBasics():
             return {'data': out,
                     'long_name': ['doubleMLTlong', 'tripleMLTlong'],
                     'units': ['hours1', 'hours2']}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert self.testInst.meta['doubleMLT'].units == 'hours1'
         assert self.testInst.meta['doubleMLT'].long_name == 'doubleMLTlong'
@@ -174,7 +173,7 @@ class TestBasics():
             out.name = 'doubleMLT'
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1'}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert self.testInst.meta['doubleMLT'].units == 'hours1'
         assert self.testInst.meta['doubleMLT'].long_name == 'doubleMLTlong'
@@ -187,7 +186,7 @@ class TestBasics():
             out.name = 'doubleMLT'
             return {'data': out,
                     'units': 'hours1'}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert self.testInst.meta['doubleMLT'].units == 'hours1'
         assert self.testInst.meta['doubleMLT'].long_name == 'doubleMLT'
@@ -199,7 +198,7 @@ class TestBasics():
                                index=inst.index)
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1', 'name': 'doubleMLT'}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert self.testInst.meta['doubleMLT'].units == 'hours1'
         assert self.testInst.meta['doubleMLT'].long_name == 'doubleMLTlong'
@@ -213,7 +212,7 @@ class TestBasics():
             # out.name = 'doubleMLT'
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1'}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
 
     def test_add_numpy_array_w_meta_name_in_dict(self):
@@ -221,7 +220,7 @@ class TestBasics():
             out = 2.*inst['mlt'].values
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1', 'name': 'doubleMLT'}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert self.testInst.meta['doubleMLT'].units == 'hours1'
         assert self.testInst.meta['doubleMLT'].long_name == 'doubleMLTlong'
@@ -233,7 +232,7 @@ class TestBasics():
             out = (inst.data.mlt * 2).values
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1'}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
 
     def test_add_list_w_meta_name_in_dict(self):
@@ -241,7 +240,7 @@ class TestBasics():
             out = (inst.data.mlt * 2).values.tolist()
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1', 'name': 'doubleMLT'}
-        self.add(custom1, 'add')
+        self.attach(custom1, 'add')
         self.testInst.load(2009, 1)
         assert self.testInst.meta['doubleMLT'].units == 'hours1'
         assert self.testInst.meta['doubleMLT'].long_name == 'doubleMLTlong'
@@ -253,7 +252,7 @@ class TestBasics():
             out = (inst.data.mlt * 2).values.tolist()
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1'}
-        self.testInst.custom.add(custom1, 'add')
+        self.testInst.custom.attach(custom1, 'add')
         self.testInst.load(2009, 1)
 
     def test_clear_functions(self):
@@ -261,7 +260,7 @@ class TestBasics():
             out = (inst.data.mlt * 2).values
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1', 'name': 'doubleMLT'}
-        self.testInst.custom.add(custom1, 'add')
+        self.testInst.custom.attach(custom1, 'add')
         self.testInst.custom.clear()
         assert self.testInst.custom._functions == []
         assert self.testInst.custom._kind == []
@@ -270,7 +269,7 @@ class TestBasics():
         def custom1(inst):
             out = (inst.data.mlt * 2).values
             return
-        self.testInst.custom.add(custom1, 'pass')
+        self.testInst.custom.attach(custom1, 'pass')
         self.testInst.load(2009, 1)
 
         assert True
@@ -281,7 +280,7 @@ class TestBasics():
             out = (inst.data.mlt * 2).values
             return {'data': out, 'long_name': 'doubleMLTlong',
                     'units': 'hours1', 'name': 'doubleMLT'}
-        self.testInst.custom.add(custom1, 'pass')
+        self.testInst.custom.attach(custom1, 'pass')
         self.testInst.load(2009, 1)
 
     @raises(AttributeError)
@@ -300,11 +299,11 @@ class TestBasics():
             out = (inst.data.tripleMLT * 2).values
             return {'data': out, 'long_name': 'quadMLTlong',
                     'units': 'hours1', 'name': 'quadMLT'}
-        self.testInst.custom.add(custom1, 'add')
-        self.testInst.custom.add(custom2, 'add')
+        self.testInst.custom.attach(custom1, 'add')
+        self.testInst.custom.attach(custom2, 'add')
         # if this runs correctly, an error will be thrown
         # since the data required by custom3 won't be present yet
-        self.testInst.custom.add(custom3, 'add', at_pos=1)
+        self.testInst.custom.attach(custom3, 'add', at_pos=1)
         self.testInst.load(2009, 1)
 
 
@@ -336,31 +335,3 @@ class ConstellationTestBasics(TestBasics):
     def add(self, function, kind='add', at_pos='end', *args, **kwargs):
         """ Add a function to the object's custom queue"""
         self.testConst.data_mod(function, kind, at_pos, *args, **kwargs)
-
-class TestDeprecation():
-
-    def setup(self):
-        """Runs before every method to create a clean testing setup"""
-        warnings.simplefilter("always")
-
-    def teardown(self):
-        """Runs after every method to clean up previous testing"""
-
-    def test_deprecation_warning_custom_add(self):
-        """Test if custom.add is deprecated"""
-
-        def func():
-            """Fake function to attach"""
-            print('Hi!')
-
-        testInst = pysat.Instrument(platform='pysat', name='testing')
-        with warnings.catch_warnings(record=True) as war:
-            try:
-                testInst.custom.add(func)
-            except AttributeError:
-                # Setting inst to None should produce a AttributeError after
-                # warning is generated
-                pass
-
-        assert len(war) >= 1
-        assert war[0].category == DeprecationWarning