+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Fixed pysat_testing method definition to include mangle_file_dates keyword
   - Added small time offsets (< 1s) to ensure COSMIC files and data have unique times
   - Updates to Travis CI environment
+  - Removed `inplace` use in xarray `assign` function, which is no longer allowed
 
 
 ## [2.1.0] - 2019-11-18